+++ EDA(trab_4)/metodo_dobra.h
@@ -1,1 +1,1 @@
-int metodo_da_dobra(int chave, int casas_tam);
+long long int metodo_da_dobra(long long int chave, int casas_tam);
